+++ CHANGELOG.md
@@ -6,14 +6,13 @@ All notable changes to OPNsense Manager will be documented in this file.
 ## [1.7.0] - 
 
 ### Added
-
 - Added Automatic neighbor discovery with pagination, the ability to search and start / stop / restart the service
 - Added Thermal Sensors information on the dashboard screen for applicable systems (Will not work with VMs)
 
 ### Changed
+- Adjusted the services section to show the description instead of the name to mimic the opnsense GUI
 
 ### Fixed
-
 - Resolved an issue on older android devices (Android 10) being unable to export profiles due to permission issues
 - Resolved a bug for KEA DHCP not displaying the proper lease type